pyproject.toml
[project]
name = "alpha-chess"
version = "0.1.0"
description = "A chess bot inspired by AlphaGo, implemented with PyTorch."
authors = [
    { name = "Cline", email = "cline@example.com" }
]
dependencies = [
    "python-chess>=1.99",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "torchrl>=0.8.1",
    "numpy>=2.3.0",
    "tqdm>=4.67.1",
    "tensorboard>=2.19.0",
    "PyYAML>=6.0.2", # For configuration
]
requires-python = ">=3.12, <3.13"
readme = "README.md"

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128" },
]
torchvision = [
  { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = ["src"]
